zipShpComponent function moved from omnivore.utils to betterfilelayer.utils
