+++ src/leaflet.omnivore.utils.js
@@ -1,5 +1,3 @@
-import JSZip from "jszip";
-
 /**
  * Download a blob object by its url and read it as a text
  *
@@ -37,31 +35,4 @@ export function parseXML(str) {
     return (new DOMParser()).parseFromString(str, 'text/xml');
   }
   return str;
-}
-
-/**
- * Gets the object returned from extractShpComponents function and create a list of
- * in-memory zipped shapefiles from every key.
- * In summary: Gets the shapefile components (.shp, .dbf, .prj ...) grouped by name and zip compress
- * @param {Object} shapes
- * @returns {Array}
- */
-export async function zipShpComponents(shapes) {
-  const zips = [];
-
-  for (const shapeName in shapes) {
-    const zip = new JSZip();
-
-    for (const component of shapes[shapeName]) {
-      zip.file(component.name, component.arrayBuffer());
-    }
-
-    const blob = await zip.generateAsync({ type: "blob" });
-
-    const shpZip = new File([blob], `${shapeName}.zip`, { type: "application/zip" });
-
-    zips.push(shpZip);
-  }
-
-  return zips;
 }
